11-20 purge api sample code

+++ src/main/java/com/gscdn/managedapi/repo/UserJpaRepo.java
@@ -1,9 +1,10 @@
 package com.gscdn.managedapi.repo;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gscdn.managedapi.model.entity.User;
 
-public interface UserJpaRepo extends JpaRepository<User, Integer> {
-
+public interface UserJpaRepo extends JpaRepository<User, Long> {
 }
